requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.33
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# AI & ML
openai==1.5.0
pandas==2.1.4
numpy==1.24.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & File handling
httpx==0.25.2
aiofiles==23.2.1
python-magic==0.4.27

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0

# Production
gunicorn==21.2.0